Fixed IPJ version, test after merging
