+++ notary-btc-integration-test/src/integration-test/kotlin/integration/btc/BtcWithdrawalFailResistanceIntegrationTest.kt
@@ -80,7 +80,7 @@ class BtcWithdrawalFailResistanceIntegrationTest {
         val randomNameSrc = String.getRandomString(9)
         val testClientSrcKeypair = ModelUtil.generateKeypair()
         val testClientSrc = "$randomNameSrc@$CLIENT_DOMAIN"
-        val res = khttp.post(
+        var res = khttp.post(
             "http://127.0.0.1:${registrationServiceEnvironment.registrationConfig.port}/users",
             data = mapOf("name" to randomNameSrc, "pubkey" to testClientSrcKeypair.public.toHexString())
         )
@@ -89,6 +89,12 @@ class BtcWithdrawalFailResistanceIntegrationTest {
             integrationHelper.registerBtcAddressNoPreGen(randomNameSrc, CLIENT_DOMAIN, testClientSrcKeypair)
         integrationHelper.sendBtc(btcAddressSrc, 1, environment.btcWithdrawalConfig.bitcoin.confidenceLevel)
         val randomNameDest = String.getRandomString(9)
+        val testClientDestKeypair = ModelUtil.generateKeypair()
+        res = khttp.post(
+            "http://127.0.0.1:${registrationServiceEnvironment.registrationConfig.port}/users",
+            data = mapOf("name" to randomNameDest, "pubkey" to testClientDestKeypair.public.toHexString())
+        )
+        assertEquals(200, res.statusCode)
         val btcAddressDest = integrationHelper.registerBtcAddressNoPreGen(randomNameDest, CLIENT_DOMAIN)
         integrationHelper.addIrohaAssetTo(testClientSrc, BTC_ASSET, amount)
         integrationHelper.transferAssetIrohaFromClient(
