+++ notary-commons/build.gradle
@@ -17,7 +17,7 @@ dependencies {
     compile "io.reactivex.rxjava2:rxkotlin:2.0.0"
 
     // Iroha Pure Java library
-    compile 'com.github.hyperledger.iroha-java:client:master-SNAPSHOT'
+    compile 'com.github.hyperledger.iroha-java:client:3.2.1'
 
     // logging
     compile 'io.github.microutils:kotlin-logging:1.4.9'
